Fix auto proxied return values (#14)
When we call a function from the wrapped side, the new message handler added in
and WireValueType. If we get a WireValue in the exposeInner handler we can
return early.

This is a bit annoying because they are const enums and perversely typescript
has no automatic way to list the values of a const enum. (This feature was
requested in

microsoft/TypeScript#21391

but the typescript devs did not seem to understand the request? In any
case, this self-evidently useful feature is marked as wont-fix.)

To work around this, I set up a way to generate a Set from a const enum with
manual duplication of the enum code. This is type checked so that if the two
blocks get out of sync at all there will be a type error.

While I was at it I added a static type assertion that WireValueType and
MessageType are disjoint.